product_management/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0,<6.0
django-filter>=22.1,<23.0
openpyxl>=3.0,<4.0
django-ratelimit>=2.1,<3.0
